+++ Tests/Wire/Resources/Shipping/OrderDrafts/Expanded/ExpandedClientWireTests.swift
@@ -414,6 +414,7 @@ import Chrt
             filterCourierOrgId: "filter_courier_org_id",
             filterForwarderOrgId: "filter_forwarder_org_id",
             filterShipperOrgId: "filter_shipper_org_id",
+            filterOffChrtShipperOrgId: "filter_off_chrt_shipper_org_id",
             request: .init(body: OrderAndTaskGroupExpandedReq(
 
             )),